+++ run_jazz_dj.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+export PYGAME_HIDE_SUPPORT_PROMPT='hide'
+
+./disc_jockey.py \
+  -n 24 \
+  -d $HOME/Desktop/Jazz \
+  -r 1.2